product_pipeline/requirements.txt
flask
flask-cors
requests
python-dotenv
Pillow
numpy
opencv-python-headless
scikit-learn
torch
torchvision
ultralytics
open_clip_torch
easyocr
